"lazydocker" snowblock
Creates a new snowblock for lazydocker [1].
> The lazier way to manage everything Docker.
> A simple terminal UI for both Docker and docker-compose.

I've used it for quite some time and the configurations have now been
tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the `config.yml` file [2] symlinked to the
corresponding XDG specific [3] and OS dependent path(s) [4].
The keybindings [5] are can not be customized yet, but some initial work
has been done [6] so maybe it'll be possible in later versions.
The colors have been adjusted to match the Nord [7] theme style.

For more details about available configurations see the
`github.com/jesseduffield/lazydocker/pkg/config` package documentation
[8].

[1]: https://github.com/jesseduffield/lazydocker
[2]: https://github.com/jesseduffield/lazydocker/blob/master/docs/Config.md
[3]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
[4]: https://github.com/jesseduffield/lazydocker/blob/master/docs/Config.md#locations
[5]: https://github.com/jesseduffield/lazydocker/blob/master/docs/keybindings/Keybindings_en.md
[6]: https://github.com/jesseduffield/lazydocker/pull/117
[7]: https://www.nordtheme.com
[8]: https://pkg.go.dev/github.com/jesseduffield/lazydocker/pkg/config

GH-283

Co-authored-by: Sven Greb <development@svengreb.de>