Broken custom colors for Git integration in `igloo` theme
Due to a recent fix in the `git-prompt.sh` script [1], regarding th
color hints when running in ZSH [2], the integration in the `igloo`
theme [3] broke the custom colors. Instead of using theme colors the
colors defined by the script took precedence, resulting in unexpected
styles.
To fix the colors the activation of `GIT_PS1_SHOWCOLORHINTS` is not
necessary anymore and has therefore been set to an empty value.

[1]: https://github.com/git/git/blob/master/contrib/completion/git-prompt.sh
[2]: https://github.com/git/git/commit/ea625a39fe5dfcaa3df0d1ee53a40008213acd59
[3]: https://github.com/arcticicestudio/igloo/blob/18b7350/snowblocks/zsh/lib/themes/igloo.zsh#L230

Co-authored-by: Sven Greb <development@svengreb.de>

GH-295