`pueue` snowblock
Created a new snowblock for `pueue` [1].

" Manage your shell commands.
" A command-line task management tool for sequential and parallel execution of long-running tasks.

I've used the project for quite some time and the configurations have
now been tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the main configuration file `pueue.yml` [2] and
`pueue_aliases.yml` file for aliases [3] which are symlinked to the
corresponding XDG specific [4] and OS dependent paths where Windows is
out-of-scope:

- macOS: `~/Library/Application Support/pueue/pueue.yml`
- Linux: `$XDG_CONFIG_HOME/pueue/pueue.yml or ~/.config/pueue/pueue.yml`

[1]: https://github.com/Nukesor/pueue
[2]: https://github.com/Nukesor/pueue/wiki/Configuration
[3]: https://github.com/Nukesor/pueue/wiki/Miscellaneous#aliases
[4]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

GH-298